Report on a venue the engine can answer
packages/reporting was converted by no step in the plan, and nothing in the tree could see that: the
guard against leftover PostgreSQL SQL scans apps/server only, and says so in its own header. The
package was 15 files failing of 20, and ten report routes answered 500 on a real venue. It is 18 of
20 now, 206 tests passing where 88 did — 114 red to green and NONE green to red, diffed by sorted
test name.

Every replacement was measured against the old engine rather than reasoned about, with PGlite
standing in for PostgreSQL and a control in the other direction for each one. The business day
itself: 28,800 instants across six zones and five cutovers, no disagreement, and with the cutover
shift dropped the same probe reports 9,012. The filed date of issue: 960 instant-and-offset pairs,
no disagreement, and 144 when the stamp is read as local time. The period bounds: 68 year-and-period
pairs, none, and 49 when a month is taken as thirty days. Per-line VAT rounding in its integer form:
220 pairs, none — and in the floating-point form, all 220 wrong.

Two of those probes were themselves wrong first, both the shape CLAUDE.md §1 names: one compared
PostgreSQL's rendered text lexicographically against ISO strings and invented 18,608 mismatches, the
other keyed a join on a rendered timestamp that the session's own time zone had re-labelled, and
invented 757. A third measurement went against its own expectation and is recorded that way: the
explicit sign in a printf'd offset makes no difference, so the comment does not claim it does.

FOUR LOSSES, named rather than glossed. A business-day window that crosses a daylight-saving change
AT the cutover hour can now differ: 8 comparisons of 288,000, all at a 02:30 cutover in
Europe/Madrid, on the two days a year that zone moves — stated in a comment where it happens. The
VAT aggregate no longer refuses an out-of-range filed amount, because the numeric casts that did
that are gone. It returns one row per breakdown element rather than one per rate, because the sum
moved into exact Decimal arithmetic: SQLite has no exact decimal type, and adding filed VAT figures
as floating point is not something a fiscal report may do. And the business-day helpers no longer
take a transaction, since the database clock and the process clock are the same clock now.

The residue guard scans a named list of roots — its own package plus reporting's src and test — with
four proofs by deletion behind it: a planted now() in each of the two new roots is named with its
file and line, a misspelt root fails twice over, and emptying the one-entry exclusion list names the
file it excludes. The controller re-ran the first of those independently, and its own first attempt
proved nothing: a bare template literal is not a sql template and the guard is not meant to see it.
Planted inside a sql template, it names packages/reporting/src/counts.ts and its line.

Two files are left failing and they are not this change's: both need the real-PostgreSQL harness
this branch removed, and 48 such files remain tree-wide. Coverage, measured with those two excluded
because Vitest 4 emits no report at all when a suite fails to collect: 99.74 / 99.35 / 100 / 99.73
against a 90/90/85/85 bar.

FIVE NEGATIVE CASES WERE PASSING VACUOUSLY and are rewritten: they crafted PostgreSQL-shaped errors
for pgErrorCode, which on this engine reads a code that is the same string for every failure, so
each returned false for the wrong reason. They now assert a refusal shape measured from the engine.
Other packages may carry the same rot; it is in the plan.

+++ packages/reporting/test/fixtures.test.ts
@@ -16,7 +16,7 @@ describe("fixtures", () => {
       lines: [{ vatRate: "21.00", lineTotal: "100.00" }],
     });
     const { rows } = await suite.db.execute<{ n: number }>(
-      sql`select count(*)::int as n from sale_lines where sale_id = ${saleId}`,
+      sql`select count(*) as n from sale_lines where sale_id = ${saleId}`,
     );
     expect(rows[0]!.n).toBe(1);
   });
